# perf: small field prover improvements