--- a/prover/maths/common/smartvectors/arithmetic_gen.go
+++ b/prover/maths/common/smartvectors/arithmetic_gen.go
@@ -1,7 +1,6 @@
 package smartvectors
 
 import (
-	"github.com/consensys/linea-monorepo/prover/maths/common/mempool"
 	"github.com/consensys/linea-monorepo/prover/maths/field"
 	"github.com/consensys/linea-monorepo/prover/utils"
 )
@@ -11,7 +10,7 @@ import (
 //   - The function panics if svecs is empty
 //   - The function panics if the length of coeffs does not match the length of
 //     svecs
-func LinComb(coeffs []int, svecs []SmartVector, p ...mempool.MemPool) SmartVector {
+func LinComb(coeffs []int, svecs []SmartVector) SmartVector {
 	// Sanity check : all svec should have the same length
 	length := svecs[0].Len()
 	for i := 0; i < len(svecs); i++ {
@@ -19,7 +18,7 @@ func LinComb(coeffs []int, svecs []SmartVector, p ...mempool.MemPool) SmartVecto
 			utils.Panic("bad size %v, expected %v", svecs[i].Len(), length)
 		}
 	}
-	return processOperator(linCombOp{}, coeffs, svecs, p...)
+	return processOperator(linCombOp{}, coeffs, svecs)
 }
 
 // Product computes a product of smart-vectors with integer exponents
@@ -27,8 +26,8 @@ func LinComb(coeffs []int, svecs []SmartVector, p ...mempool.MemPool) SmartVecto
 //   - The function panics if svecs is empty
 //   - The function panics if the length of exponents does not match the length of
 //     svecs
-func Product(exponents []int, svecs []SmartVector, p ...mempool.MemPool) SmartVector {
-	return processOperator(productOp{}, exponents, svecs, p...)
+func Product(exponents []int, svecs []SmartVector) SmartVector {
+	return processOperator(productOp{}, exponents, svecs)
 }
 
 // processOperator computes the result of an [operator] and put the result into res
@@ -36,7 +35,7 @@ func Product(exponents []int, svecs []SmartVector, p ...mempool.MemPool) SmartVe
 //   - The function panics if svecs is empty
 //   - The function panics if the length of coeffs does not match the length of
 //     svecs
-func processOperator(op operator, coeffs []int, svecs []SmartVector, p ...mempool.MemPool) SmartVector {
+func processOperator(op operator, coeffs []int, svecs []SmartVector) SmartVector {
 
 	// There should be as many coeffs than there are vectors
 	if len(coeffs) != len(svecs) {
@@ -100,7 +99,7 @@ func processOperator(op operator, coeffs []int, svecs []SmartVector, p ...mempoo
 	}
 
 	// Accumulate the regular part of the vector
-	regularRes, matchedRegular := processRegularOnly(op, svecs, coeffs, p...)
+	regularRes, matchedRegular := processRegularOnly(op, svecs, coeffs)
 
 	// Sanity-check : all of the vector should fall into only one of the two
 	// category.
@@ -114,13 +113,13 @@ func processOperator(op operator, coeffs []int, svecs []SmartVector, p ...mempoo
 	case matchedRegular+matchedConst == totalToMatch:
 		// In this case, there are no windowed in the list. This means we only
 		// need to merge the const one into the regular one before returning
-		op.constTermIntoVec(regularRes.Regular, &constRes.Value)
+		op.constTermIntoVec(*regularRes.(*Regular), &constRes.Value)
 		return regularRes
 	default:
 
 		// If windowRes is a regular (can happen if all windows arguments cover the full circle)
 		if w, ok := windowRes.(*Regular); ok {
-			op.vecTermIntoVec(regularRes.Regular, *w)
+			op.vecTermIntoVec(*regularRes.(*Regular), *w)
 			return regularRes
 		}
 
@@ -130,7 +129,7 @@ func processOperator(op operator, coeffs []int, svecs []SmartVector, p ...mempoo
 		// In this case, the constant is already accumulated into the windowed.
 		// Thus, we just have to merge the windowed one into the regular one.
 		interval := windowRes.interval()
-		regvec := regularRes.Regular
+		regvec := *regularRes.(*Regular)
 		length := len(regvec)
 
 		// The windows rolls over