… prover/crypto/vortex/reedsolomon_ext_test.go
… prover/crypto/vortex/reedsolomon_test.go
… prover/maths/common/mempool/debug_pool.go
… prover/maths/common/mempool/debug_pool_test.go
… prover/maths/common/mempool/from_sync_pool.go
… prover/maths/common/mempool/mempool.go
… prover/maths/common/mempool/slice_arena.go
… prover/maths/common/smartvectors/arithmetic_ext_test.go
… prover/maths/common/smartvectors/arithmetic_gen_ext.go
… prover/maths/common/smartvectors/arithmetic_test.go
… prover/maths/common/smartvectors/fft.go
… prover/maths/common/smartvectors/fft_ext.go
… prover/maths/common/smartvectors/fft_ext_test.go
… prover/maths/common/smartvectors/fft_test.go
… prover/maths/common/smartvectors/regular.go
… prover/maths/common/smartvectors/regular_ext.go
… prover/maths/common/smartvectors/rotated.go
… prover/maths/common/smartvectors/rotated_ext.go
… prover/maths/common/smartvectors/smartvectors.go
… prover/maths/common/smartvectors_mixed/arithmetic_basic.go
… prover/maths/common/smartvectors_mixed/arithmetic_gen.go
… prover/maths/field/fext/utils.go
… prover/protocol/compiler/globalcs/quotient.go
… prover/protocol/compiler/horner/horner.go
… prover/protocol/compiler/mpts/multipoint_to_singlepoint_test.go
… prover/protocol/compiler/mpts/prover.go
… prover/protocol/compiler/splitextension/split.go
… prover/protocol/compiler/standard_benchmark_test.go
… prover/protocol/dedicated/reedsolomon/reedsolomon.go
… prover/protocol/dedicated/reedsolomon/reedsolomon_test.go