This PR ends up touching a lot of packages, mostly better memory patterns . Not a very clean PR per say, but want to merge before having too many conflicts with other branches.

Killed the `mempool` package.

```
benchmark                                  old ns/op       new ns/op      delta
BenchmarkCompiler/realistic-segment-96     11865214766     8563912192     -27.82%

benchmark                                  old allocs     new allocs     delta
BenchmarkCompiler/realistic-segment-96     112689417      110565034      -1.89%

benchmark                                  old bytes        new bytes       delta
BenchmarkCompiler/realistic-segment-96     143645704856     85597619136     -40.41%
```